Logrotate formula
The logrotate utility is designed to simplify the administration of log files
on a system which generates a lot of log files. Logrotate allows for the
automatic rotation compression, removal and mailing of log files. Logrotate
can be set to handle a log file daily, weekly, monthly or when the log file
gets to a certain size. Normally, logrotate runs as a daily cron job.

The package is not buildable on Darwin without including an additional
header, nor is it installable without teaching the Makefile what INSTALL
is on Darwin.

Signed-off-by: Max Howell <max@methylblue.com>

I made a number of modifications, including comma separating the make call so
that the prefix doesn't need to be escaped if it has spaces in it.

Also setting LDFLAGS to /usr/local/lib is done in brewkit.rb already. The
POPT_DIR var shouldn't be necessary either, but I haven't tested that yet.