Exuberant Ctags
Ctags comes with OS X but Exuberant is better.

Ctags generates an index (or tag) file of language objects found in source
files that allows these items to be quickly and easily located by a text
editor or other utility. A tag signifies a language object for which an index
entry is available (or, alternatively, the index entry created for that
object).