GNU Go formula
GNU Go is a free program that plays the game of Go. GNU Go has played thousands
of games on the NNGS Go server. GNU Go is now also playing regularly on the Legend
Go Server in Taiwan, on the WING server in Japan, and many volunteers run GNU Go
clients on KGS. GNU Go has established itself as the leading non-commercial go program
in the recent tournaments that it has taken part in.