Added pip and setuptools formula
As discussed on IRC this simplifies the Python
packaging strategy of Homebrew. It's a modified
version of the hints given on http://bit.ly/HomebrewPython
and renders them unnecessary.

setuptools and pip will be installed separately
from Python's site-packages (e.g.
"#{prefix}/lib/pip") but will install other
packages in the Python's site-packages directory
and *not* inside their own directories.

In case the Python brew is removed or unlinked
all packages installed with pip will be removed
or unlinked, too. When using the non-framework
install of Python (e.g. not the OS X' default)
these restrictions also apply in case of the
installation of a new version of Python.