Libtasn 2.3
The ASN.1 library used by GnuTLS, GNU Shishi and some other packages.

It was written by Fabio Fiorina, and has been shipped as part
of GnuTLS for some time but is now a proper GNU package.

The goal of this implementation is to be highly portable, and only
require an ANSI C89 platform.