added formula for install Sproxy
Sproxy is a complimentary tool for the load testing tool Siege. Sproxy is the
[S]iege[proxy], a tool designed to harvest URLs to a urls.txt for the Siege
http/https regression tester. Sproxy is designed to work like a regular http
proxy. It listens for http requests on port 9001 (default) and it returns
requested content to the web browser while it writes the URL along with any
GET requests and/or POST data to a file for Siege.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>

* Fix man file location.