Update SML-NJ
The existing SML-NJ formula does all sorts of nasty things, like spewing
ML files into lib and polluting bin with hidden dot-folders.

Instead of removing it just yet, making it install into libexec.

If you can improve this formula, go for it, otherwise it might be retracted
in the future.