ChucK 1.2.1.3
A new (and developing) audio programming language for real-time synthesis,
composition, performance, and now, analysis - fully supported on MacOS X,
Windows, and Linux.

ChucK presents a new time-based, concurrent programming model that's highly
precise and expressive (we call this strongly-timed), as well as dynamic
control rates, and the ability to add and modify code on-the-fly.

In addition, ChucK supports MIDI, OSC, HID device, and multi-channel audio.
It's fun and easy to learn, and offers composers, researchers, and performers
a powerful programming tool for building and experimenting with complex audio
synthesis/analysis programs, and real-time interactive control.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>