add aardvark_shell_utils
Aardvark shell utils includes three helpful programs:

realpath(1) is an implementation of the realpath(3) function for returning the
canonicalized absolute pathname of an input path.
filebase(1) returns the base portion of a filename, that is the filename with
its extension (e.g. .txt) removed.
fileext(1) returns the extension portion of a filename (e.g. txt).

Signed-off-by: Adam Vandenberg <flangy@gmail.com>