Add formula for the Judy dynamic array library
From the website:

"Judy is a C library that provides a state-of-the-art core technology
that implements a sparse dynamic array. Judy arrays are declared simply
with a null pointer. A Judy array consumes memory only when it is
populated, yet can grow to take advantage of all available memory if
desired."

Signed-off-by: Mike McQuaid <mike@mikemcquaid.com>