Asymptote: Fix installation
Asymptote needs a serialized `make install` or else things fail due to race
conditions.

Also, altered installation so that TeX packages are no long installed to texmf
trees outside of the Asymptote keg---notably:

    /usr/local/texlive/<dist_year>/texmf-local
    ~/Library/texmf

All packages are now installed to HOMEBREW_PREFIX/share/texmf and instructions
for adding this directory to the TeX search path are given in the caveats.