libmikmod: 3.2.0b2, use git build head only
libmikmod built from tarball is 3.2.0-beta2.
libmikmod built from cvs is called 3.2.0b2 in their code.

The final work on libmikmod was left in cvs and includes numerous
bug fixes and improvements done over the course of a year after
the tarball came out.  This adds the option to build head,
the final version.  No work has been done in five years.

Patches from Macports were merged upstream during that time.
One patch was not and is still in this formula.  Another
patch is included against the cvs version to fix a duplicate
symbol problem with vc_callback().

The 2005 era code is in need of an autoreconf -ivf, which is
done after setting an environment variable to pickup glibtoolize.
The build is done out of source according to the directions.

This builds static and shared native libraries using llvm-2335.9,
gcc-4.2.1(v5666), and clang-2.0, tested on 64bit OSX Snow Leopard.
It also builds without error on Lion.

EDIT: Builds from clone of cvs in mistydemeo with inline patches

Closes Homebrew/homebrew#8274.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>