git: define NO_R_TO_GCC_LINKER
Inadvertently failing to pass the correct library paths to the linker
can cause a build to link against an incorrect version of a library.
Some compilers don't pass the runtime path defined with '-Rpath' to the
linker. With this defined, paths will be passed to the linker with
-Wl,rpath, rather than -R.

Apple's gcc-4.0, gcc-4.2, and llvm-gcc do not complain about -R, but it
is not documented as far as I can tell, and clang complains vocally
about it. As long as they all support the -Wl format, we may as well be
safe.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>