GFortran: Re-organize and fix installation logic
Check for XCode >= 4.2 on Lion first before casing the GCC build in the install
section as this is the order in which the  download section selects which Lion
binaries to fetch.

Without this change, a leftover `gcc-4.2` installation can fool the logic into
thinking the 5666-3 binaries were downloaded instead of the full GCC-4.2 suite.
This causes the wrong installation steps to be executed which leaves users
without some important symlinks.

Closes Homebrew/homebrew#9398.
Fixes Homebrew/homebrew#9395.
Fixes Homebrew/homebrew#9043.