New Formula: movgrab
Movgrab is a command line program to download videos from the net.
Given a url, it has the code to handle about 25 different sites.
This version 1.1.9 compiles on Lion and 64bit Snow Leopard.
It will occasionally segfault while running the CLI during a
download, but Colum, the author, has not been able to track it
down as of yet because it is not consistent.  He is lacking a Mac
to test it using gdb and could use help from anyone with experience
in network steams, C, and gdb.

EDIT: moved patches to a DATA block rather than inreplace
EDIT: fix the "inline" issue with c99 & Clang.
EDIT: update from 1.1.8 to 1.1.9.  No changes to patches.
EDIT: all patches including --no-recursion are upstream via email.

This would close Homebrew/homebrew#10074

Closes Homebrew/homebrew#10101.

Signed-off-by: Misty De Meo <mistydemeo@gmail.com>