Use Homebrew's pcre where appropriate
We provide a pcre formula as OS X does not provide libpcre headers.
However, it does provide the library itself and so occasionally things
will use Homebrew's pcre headers and link against the system library.
This should be avoided to prevent incompatibilities, so explicitly tell
the build system where to look for the library where necessary.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>