fping: Update to 3.1
New website, and new maintainer.

fping has been mothballed since 2002.  A new maintainer has taken
over this project, set up a new website, fping.org, and committed
the project on GitHub.

Change homepage.
Upgrade to version 3.1
Add head url on GitHub
Remove `--man` because it uses a modern configure that works.
Add caveat about this software needing to be owned by root
and run as root or with the setuid bit set.  An issue is open
with the developer to see if there is a workaround for that.

Closes Homebrew/homebrew#11931.

Signed-off-by: Charlie Sharpsteen <source@sharpsteen.net>