qt: add --developer-build option
Qt4 has a configure option, `--developer-build` that compiles and
links Qt with Qt developer options.  This feature was requested
in an issue.  Add that as a user install option.  This has no
affect on the bottles, which can remain unchanged.  This compiles
without error on Lion using llvm-gcc from XCode-4.3.2.

Closes Homebrew/homebrew#12284.
Closes Homebrew/homebrew#12595.

Signed-off-by: Charlie Sharpsteen <source@sharpsteen.net>