Cntlm 0.92.3
Cntlm is an NTLM authenticating HTTP proxy which is very useful if you
work in a company with a Windows based proxy for outgoing Internet
traffic. It sits between your applications and the corporate proxy,
adding NTLM authentication on-the-fly.

Compared to alternatives like NTLMAPS, Cntlm is small, fast and has no
dependencies.

Closes Homebrew/homebrew#14004.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>