cairo: always depend on our pixman
Older versions of XQuartz, including the one distributed by Apple with
OS X, contain an insufficient version of pixman for recent cairo.

Potentially the :pixman dep could be taught to only delegate to XQuartz
if it is up to date enough, but until then, always depend on our pixman.

Fixes Homebrew/homebrew#14072.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>