gst-plugins-ugly: fix compile error, includes
gst-plugins-ugly builds against opencore-amr and libcdio.
Add those optional deps to be specific and consistent.
gst-plugins-ugly fails to compile when `opencore-am?b` are
installed due to missing includes.  Add the location of them
to the appropriate env vars.  Tested on ML with clang and llvm
from XCode-4.4.1.

Fixes Homebrew/homebrew#14078

Closes Homebrew/homebrew#14085.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>