Heroku Toolbelt 2.32.13
Heroku has deprecated their Rubygem, and is encouraging everyone to use
their Toolbelt to install and use the `heroku` command line tool. While
they have a GUI installer for OS X, that's just not how I roll when it
comes to CLI tools. This takes the underlying package that the Toolbelt
is based on and wraps it up in a formula everyone can use.

Closes Homebrew/homebrew#15480.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>