Clean up lib path searching for Spatialite
Anything linking against SpatiaLite also needs to use Homebrew's SQLite so that
loadable extensions are enabled along with RTree indices.

This commit cleans up several formulae to use the specific `opt/sqlite/lib`
prefix rather than `HOMEBREW_PREFIX/lib` and also adds this searching to
`gdal`.

However, this hack is getting a little large. The best fix would probably be to
flag SQLite as `keg_only` so that lib path munging is handled automagically for
everything with a dependency tree that includes SQLite.