boost: c++11 support.
This commit modifies boost.rb to provide C++11 support for compiling Boost.  The support is the simplest possible, adding a "with-c++11" option which if selected will add the appropriate toolset switch to the bootstrap.sh invocation and the appropriate toolset, cxxflags and linkflags to the b2 invocation.  Due to a bug in Boost, it is also necessary to provide a patch (IAW Boost Ticket 7671) or C++11 compilation will fail.

Closes Homebrew/homebrew#16078.

Signed-off-by: Mike McQuaid <mike@mikemcquaid.com>