python: fix installing on NFS
python has a severe bug which prevents it from building, or performing
certain filesystem operations, on NFS.

A few functions in the shutil module, notably chflag(), will attempt
to copy file metadata which includes flags. Certain filesystems, such
as NFS, don't support copying flags and will raise ENOTSUP when it is
attempted. Unfortunately a portion of the python build process
indirectly uses chflag() to copy metadata, which results in a build
failure if the install prefix is on an NFS volume.

This was patched in 2010, but the patch didn't actually work on OS X.
This upstream patch fixes it for real.