vit: Fix installation of vit-commands help file
vit-commands is installed into 'etc', but it needs to go in
`$prefix/etc`, not in `$HOMEBREW_PREFIX/etc`. It's not actually a
configuration file, and vit looks for it in its keg prefix.

Ideally the file wouldn't even be symlinked into `$HOMEBREW_PREFIX/etc`
but I don't see any way to avoid that.

Closes Homebrew/homebrew#30886.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>