ffmpeg: use openjpeg instead of jpeg2000.
Even if -libopenjpeg is installed ffmpeg will try to use the built-in jpeg2000 codec to open j2K encoded video, which mostly doesn't work. I added a line to the build args to turn off the dodgy default decoder if someone installs --with-libopenjpeg because they probably want a working j2k codec.

Closes Homebrew/homebrew#31182.

Signed-off-by: Mike McQuaid <mike@mikemcquaid.com>