cassandra 2.1.4
Fixes CVE-2015-022.

This was intended to just be a bump to fix the CVE, but then I noticed
the caveats and thought we could do better than just shoving users onto
pip. That might have been a bad idea, since it led to me rewriting a
good chunk of the formula. I’ll ping Tim for some review of the Python
changes, whether he thinks it’s worthwhile, etc.

Closes Homebrew/homebrew#38309.

Signed-off-by: Tim D. Smith <git@tim-smith.us>