luarocks: migrate into lua formulae
This PR ‘pythonises’ the Lua setup. There have been requests for us to
do this for a while, and it’s a good idea - It’s just taken me a while
to finish ironing out.

The individual luarocks formula dies entirely, in favour of luarocks
being vendored into the two Lua formulae in Homebrew/homebrew with
versioned trees.

This provides roughly the same experience we deliver with pip
and co from Python.

Luajit is not yet supported, because it’s being a pain, but Luajit
support will land in the semi-near future, as will support for Lua53
in Versions. All 4 installations can sit side-by-side with no non-binary
conflict. Binary conflict is handled roughly the same way Python
handles it sadly, which is regrettable but presently unavoidable.