md5sha1sum: explicitly set openssl paths
configure has a hardcoded set of include and library paths, and fails
if it can't find openssl in any of them. This causes it to pick up the
wrong openssl on older OS Xs, and to fail to pick up any on 10.11.

It actually links against the correct openssl on older OS Xs, so the
revision isn't being bumped and bottles aren't being rebuilt.

Closes Homebrew/homebrew#42912.

Signed-off-by: Misty De Meo <mistydemeo@gmail.com>