Merge the fsharp recipe into the mono recipe
For reasons related to the way the mono runtime works, the installations
must be "on top" of each other for the MSBuild compilation
infrastructure to function properly. This change matches the official
Mono installer.

The fsharp installation is now optional (but defaults to on). This new
formula overwrites old fsharp links.

Fixes Homebrew/homebrew#36286

Closes Homebrew/homebrew#43719.

Signed-off-by: Xu Cheng <xucheng@me.com>