embulk 0.7.10
Update embulk to 0.7.10

Fix to execute embulk command as a shell script rather than jar

bin/embulk is a polygot script, and it also works as a shell script.
bin/embulk supports serveral options as a shell script although jar
execution does not support. So, we should run embulk as a command.

Closes Homebrew/homebrew#46983.

Signed-off-by: Dominyk Tiller <dominyktiller@gmail.com>