nginx: stop permitting userpaths
The history on this formula is rather long but as far as I can trace userpaths
was needed because for a significant amount of time we permitted the passenger
option to pluck out the first `passenger-config` in the $PATH and use that.

That changed about two years ago now in https://github.com/Homebrew/homebrew-core/commit/d5719a8135ca0b01bef0bd100d28f33f7d1f5151
and userpaths should have probably been removed at that point, but since it wasn't
explicitly breaking anything it went unnoticed and hasn't been discussed since.

Nothing about the remaining formula suggests userpaths should still be needed and
it doesn't break if you remove that element, so let's remove it.